Base: Allow hiding notification icons
When long pressing a notification, add another menu
item which allows the user to hide the notification icon
(but keep the actual notification in the dropdown).

This is meant for applications which have services running
in the statusbar which the user is aware of, so right now
we hide the icon by the package.

Change-Id: I60c8705e1fd02d7b26dbe90fe39a3b40febdbfcc
Signed-off-by: anerikg <anerikgames@gmail.com>

Conflicts:
	packages/SystemUI/res/values/cr_strings.xml
	packages/SystemUI/src/com/android/systemui/statusbar/BaseStatusBar.java

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/BaseStatusBar.java

Conflicts:
	packages/SystemUI/res/values/slim_strings.xml